Moving tract-page into the page folder
